+++ src/main/java/resources/Quotes.json
@@ -2,32 +2,32 @@
   {
     "type" : "SINGLE",
     "name" : "mainHelp",
-    "output" : "Это учебный бот!!! Пока тут есть только тесты для Джаваскрипта и Английского языка\nКоманды:\nJS - для перехода в меню обучения ДжаваСкрипту\nENG - для перехода в меню обучения Английскому\nexit - чтобы прервать работу программы"
+    "response" : "Это учебный бот!!! Пока тут есть только тесты для Джаваскрипта и Английского языка\nКоманды:\nJS - для перехода в меню обучения ДжаваСкрипту\nENG - для перехода в меню обучения Английскому\nexit - чтобы прервать работу программы"
   },
   {
     "type" : "SINGLE",
     "name" : "JSHelp",
-    "output" : "Это меню обучения ДжаваСкрипту\nКоманды:\nquest - чтобы получить вопрос с вариантами ответов\nexit - чтобы вернуться в главное меню"
+    "response" : "Это меню обучения ДжаваСкрипту\nКоманды:\nquest - чтобы получить вопрос с вариантами ответов\nexit - чтобы вернуться в главное меню"
   },
   {
     "type" : "SINGLE",
     "name" : "toMenu",
-    "output" : "Назад в меню!"
+    "response" : "Назад в меню!"
   },
   {
     "type" : "SINGLE",
     "name" : "travelToJS",
-    "output" : "Переходим к изучению JavaScript!\nВводи question для получения вопроса, ну или help, для всякого другого"
+    "response" : "Переходим к изучению JavaScript!\nВводи question для получения вопроса, ну или help, для всякого другого"
   },
   {
     "type" : "SINGLE",
     "name" : "toMenu",
-    "output" : "Назад в меню!"
+    "response" : "Назад в меню!"
   },
   {
     "type" : "RANDOM",
     "name" : "unknownCommand",
-    "outputArray" : [
+    "responseRandom" : [
       "Я же сказал хелп введи, если не шаришь",
       "Я же сказал хелп ввести",
       "Нет такой команды",
@@ -40,7 +40,7 @@
   {
     "type" : "RANDOM",
     "name" : "correctAnswer",
-    "outputArray" : [
+    "responseRandom" : [
       "Верно!",
       "Правильный ответ!",
       "Верно. Наш слоняра",
@@ -60,7 +60,7 @@
   {
     "type" : "RANDOM",
     "name" : "incorrectAnswer",
-    "outputArray" : [
+    "responseRandom" : [
       "...неверно",
       "Ну почти...",
       "Вообще ни разу",
