+++ src/main/java/structures/Quote.java
@@ -3,15 +3,15 @@ package structures;
 public class Quote {
     public QuoteType type;
     public String name = "";
-    public String output = "";
-    public String[] outputArray = {}; //sealed class
+    public String response = "";
+    public String[] responseRandom = {}; //sealed class
 
     public Quote() {}
 
-    public Quote(String type, String name, String output, String[] outputArray) {
+    public Quote(String type, String name, String response, String[] responseRandom) {
         this.type = QuoteType.valueOf(type);
         this.name = name;
-        this.output = output;
-        this.outputArray = outputArray;
+        this.response = response;
+        this.responseRandom = responseRandom;
     }
 }
